requirements.txt

flask==3.0.0
vonage==4.7.1
requests==2.31.0
pandas==2.3.3
python-dotenv==1.1.1
pytz==2025.2

# Phase 5: Foundation & Essential Intelligence
redis==5.0.1
twilio==9.0.4
textblob==0.17.1
cryptography==41.0.7
prometheus-client==0.19.0

# Phase 6: Advanced Booking Intelligence & Customer Experience
sendgrid==6.10.0
apscheduler==3.10.4
python-dateutil==2.8.2
numpy==1.26.4
scikit-learn==1.3.0
icalendar==5.0.11

# Legacy Google Calendar dependencies (optional - only needed if reverting to Google Calendar)
# google-api-python-client==2.183.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
# python-dateutil provides relativedelta
